refactor various barcode files
